Resolve crash caused by device drivers returning invalid values for known attribute types.
Specifically, this device driver for "battery" reporting may return the string "unknown" instead of a number. https://raw.githubusercontent.com/ymerj/HE-HA-control/main/genericComponentBattery.groovy
This fix corrects a crash that would occur when a type cannot be deserialized/converted to the expected attribute type for known device attributes. Rather than crashing, the error is logged along with some information helpful to diagnose and correct what caused it. The attribute is then ignored, though hubitat2prom will continue to ingest and try to report on the attribute for subsequent requests.

See this thread for more information. https://community.hubitat.com/t/release-prometheus-device-metrics-hubitat2prom-in-c/90365/4

Closes #2